Move function to FormSheetUtils
