+++ src/components/gamma/modals/form-sheet/FormSheet.tsx
@@ -2,21 +2,10 @@ import React from 'react';
 import { StyleSheet } from 'react-native';
 import FormSheetHostNativeComponent from '../../../../fabric/gamma/modals/form-sheet/FormSheetHostNativeComponent';
 import type { FormSheetProps } from './FormSheet.types';
-import { resolveLargestUndimmedDetentIndex } from './FormSheetUtils';
-
-// TODO: @t0maboro - move to SheetUtils after merging PR with largestUndimmedDetentIndex
-export function resolveNativeCornerRadius(
-  radius?: number | 'systemDefault',
-): number | undefined {
-  if (radius === 'systemDefault') {
-    return -1.0;
-  }
-  if (typeof radius === 'number' && radius < 0) {
-    return -1.0;
-  }
-
-  return radius;
-}
+import {
+  resolveLargestUndimmedDetentIndex,
+  resolveNativeCornerRadius,
+} from './FormSheetUtils';
 
 export function FormSheet(props: FormSheetProps) {
   const {
